AF-18: PHP 8.2 Kompatibilität

See merge request asteriosframework/core!45

+++ src/Router.php
@@ -345,7 +345,7 @@ class Router
                 $params = array_map(static function ($match, $index) use ($matches) {
                     if (isset($matches[$index + 1][0]) && is_array($matches[$index + 1][0]))
                     {
-                        return trim(substr($match[0][0], 0, $matches[$index + 1][0][1] - $match[0][1]), '/');
+                        return trim(substr($match[0][0], 0, (int) ($matches[$index + 1][0][1] - $match[0][1])), '/');
                     }
 
                     return isset($match[0][0]) ? trim($match[0][0], '/') : null;
